requirements.txt
backcall==0.2.0
colorama==0.4.4
commonmark==0.9.1
cycler==0.11.0
decorator==5.1.0
et-xmlfile==1.1.0
ipdb==0.13.9
ipython==7.29.0
jedi==0.18.1
kiwisolver==1.3.2
lxml==4.6.4
matplotlib==3.4.3
matplotlib-inline==0.1.3
numpy==1.21.4
openpyxl==3.0.9
pandas==1.3.4
parso==0.8.2
pathlib==1.0.1
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.22
Pygments==2.10.0
pyparsing==3.0.5
python-dateutil==2.8.2
python-pptx==0.6.21
pytz==2021.3
rich==10.13.0
scipy==1.7.2
seaborn==0.11.2
six==1.16.0
toml==0.10.2
traitlets==5.1.1
wcwidth==0.2.5
XlsxWriter==3.0.2
